Creating breakpoints & clean up the code little bit
