+++ code_base/web-ui/src/app/pages/WorkInProgress/components/Footer.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Container, Typography, Grid, Link } from "@material-ui/core";
 import { makeStyles } from "@material-ui/core/styles";
-import { toAbsoluteUrl } from "../../../../utils/utils";
+import { toAbsoluteUrl } from "../../../utils/utils";
 
 const useStyles = makeStyles({
   root: {
