+++ code_base/web-ui/src/app/pages/WorkInProgress/components/Nav.js
@@ -2,7 +2,7 @@ import React from "react";
 import { makeStyles, useTheme } from "@material-ui/core/styles";
 import { AppBar, Container, Toolbar } from "@material-ui/core";
 import useMediaQuery from "@material-ui/core/useMediaQuery";
-import { toAbsoluteUrl } from "../../../../utils/utils";
+import { toAbsoluteUrl } from "../../../utils/utils";
 
 const useStyles = makeStyles((theme) => ({
   root: {
