--- code_base/web-ui/src/app/pages/WorkInProgress/components/mobile/Mobile.js
@@ -1,16 +1,0 @@
-import React from "react";
-import Nav from "./Nav";
-import Middle from "./Middle";
-import AboutUs from "./AboutUs";
-import Footer from "./Footer";
-
-export default function Mobile() {
-  return (
-    <>
-      <Nav />
-      <Middle />
-      <AboutUs />
-      <Footer />
-    </>
-  );
-}
